Make the Rounds and Labs queryset reusable
